chore: Update npm dependencies and add @types/lodash and defu packages
This commit updates the npm dependencies in the `package.json` file. The `@faker-js/faker` dependency is updated to version `8.4.1`, the `currency.js` dependency is updated to version `2.0.4`, the `destr` dependency is added with version `2.0.3`, the `effect` dependency is updated to version `3.5.7`, the `express` dependency is updated to version `4.19.2`, the `nanoid` dependency is updated to version `5.0.7`, the `pino` dependency is updated to version `9.3.2`, the `pino-pretty` dependency is updated to version `11.2.2`, the `rxjs` dependency is updated to version `7.8.1`, the `ws` dependency is updated to version `8.18.0`, and the `zod` dependency is added with version `3.23.8`. Additionally, the `@types/lodash` package is added with version `4.17.7` and the `defu` package is added with version `6.1.4`. This update enhances the codebase by ensuring that the project is using the latest stable versions of the dependencies and adding necessary type definitions for the lodash library and the defu package.