feat: Update onConclude method in LuckyBingoRound to return a number
This commit updates the `onConclude` method in the `LuckyBingoRound` class to return a number instead of the default `OnConclude` type. It generates a random result for the round by using the `Math.random()` function and multiplying it by 100. The result is then rounded up using the `Math.ceil()` function. This modification improves the functionality of the `LuckyBingoRound` class and provides a more accurate and reliable way to generate random results for lucky bingo rounds in the gemo library.