refactor: Update logger implementation to use consola library
This commit refactors the logger implementation in the `logger.ts` file. It replaces the `pino` logger with the `consola` library by importing the `createConsola` function. The `logger` constant is updated to use the `createConsola` function with a log level of 5. This change improves the logging functionality in the gemo library by utilizing the features provided by the `consola` library.