feat: Add control option to TimerEngineOptions
This commit adds a new control option to the TimerEngineOptions interface. The control option allows for different types of timer engines, such as single or continuous. The type can be specified as 'single' or 'continuous', and an optional delay can be provided for the continuous type. This addition enhances the flexibility and functionality of the TimerEngine class in managing game engines in the gemo library.