feat: Add Promisable type to Engine start method
This commit updates the `Engine` interface in the `engine.ts` file to use the `Promisable` type for the return value of the `start` method. This change allows for more flexibility in the implementation of the `start` method, as it can now return a promise or a non-promise value. This modification enhances the codebase by providing a more expressive and versatile contract for game engines in the gemo library.