feat: Add createMessage utility function for generating messages
This commit adds a new utility function called `createMessage` to the `gemo/utils` module. The function takes in a code number and a data object and returns a string representation of the JSON object with the code and data merged. This utility function can be used to generate messages in the application.