feat: Add StartGameCommand for initiating game rounds
This commit adds the StartGameCommand class in the src/commands/start-game-command.ts file. The StartGameCommand is responsible for initiating game rounds in the application. When executed, it triggers the `run` method of the `round` object in the specified `room`. This addition enhances the functionality of the application by allowing the start of game rounds through the StartGameCommand.